Move swarm type check
